Ignore a bunch of files and directories in $HOME
Because this uses a bare repository with $HOME as the working tree, the
repoistory is configured to not show untracked files by default. This is
to prevent commands like `git status` from showing the entirety of the
$HOME directory.

At times, however, it is valuable to include untracked files in commands
like `git status`. The files in this directory are decidedly not needed
to be included in this full view.