vim: remove vim-rfactory plugin
Simply, I never use this and instead use standard file navigation. We
generally use a single factory file, which makes this a snap. Some of
the extra features, like jumping to a specific trait, aren't worth
having an additional plugin and minimal testing of this showed it didn't
work all that well jumping to the correct factory definition.