fish: extend git to handle dotfiles for $HOME
The bare repository used for these dotfiles in the `$HOME` directory
requires additional options be based to `git`. The `dotfiles` function
wraps git to automatically apply these options and allows for `dotfiles`
to be used anywhere and still apply to the dotfiles git repository.
Attempting to use the regular `git`command in the home directory,
without the added options would error. This overwrites the `git` command
to be a function that wraps `git` and makes it $HOME directory aware and
able to handle the dotfiles git repository. Keeping `dotfiles` for now
as it's usable from any directory which is occasionally convenient.