git: remove shortcut aliases
Giving this a shot. The main motivator is when working on someone else's
machine while pairing, and having to fight muscle memory for using these
shortcuts. Assuming speed stays generally consistent, I'd rather break
the habit of using shortcuts for common git commands, in favor of the
standard commands, available on any machine.