kitty: Hide application title bar
The minimal application window somehow makes it easier to focus on
what's in the window. Probably because it more closely resembles
being in full screen mode.

Setting this to `titlebar-only` removes the title bar, but keeps the
window borders, which maintains the standard rounded-corner look and
feel of a macOS window.