weechat: Add WeeChat configuration from v3
WeeChat is chat client with a text-based UI, and that supports multiple
protocols. My use is primarily to connect to programming-related IRC
channels for questions. This copies the configuration files wholesale
from v3 of my dotfiles as it's configured in a way that worked for me
and was visually appealing.