(vim) Highlight end of changed text
The default behavior when using the change command is to remove the
changed text. Adding the `$` option to `ctoptions` instead add a "$"
marker to the end of the affected text.  This marker gives add clarity
that the desired text is being change. The changed text will be
overwritten once the new text is typed.