(git) Reconcile divergent branches with ff-only
By default, Git does not create an extra merge commit when merging a
commit that is a descendant of the current commit. Instead, the tip of
the current branch is fast-forwarded.

In most cases, fast-forward merges is what we want so we set `pull.ff`
to `only` as the default. This setting applies (and overrides, if set)
the `merge.ff` setting and can be overridden by passing the `--rebase`
or `--no-rebase` option.