(vim) Hide buffers when they are abandoned
By default vim is configured to unload a buffer when it is abandoned,
which can be frustrating when you have unwritten changes and want to
open another buffer.

Setting `hidden` instructs vim to hide the buffer instead, which means
you can have unwritten changes in a buffer and still open a new buffer
without being forced to first write or under the changes.

Also, undo buffers and marks are preserved with the buffer is open.