(vim) Stop centering window around matching line
Loupe makes it easier to see the currently selected pattern match by
applying a different `:highlight` group and by keeping the matching line
centered within the window. There is enough distinction with the
different `:highlight` group to see the current match; centering the
window is not needed and mildly annoying.