(vim) Round indent to multiple of `shiftwidth`
When set moves to the next multiple value of `shiftwidth` instead of
simply moving by `shiftwidth` characters.

For example, if `shiftwidth` 4 and the input starts at column 3, indenting
would move the text to column 5 when `shiftround` is set and would move
the text to column 7 when `shiftround` is not set.