(git) Configure default hooks
This defines the location for a set of hooks to be run across all
repositories, and avoids having to configure the hooks on a
per-repository basis.

It is more flexible that `init.templateDir`, which defines the location
for a template directory, presumably with hooks defined, to copy from
when initializing a repository.