(rspec) Default to running specs in random order
RSpec executes groups and examples in the order they are defined as the
spec files are loaded, with the caveat that each group runs its examples
before running its nested example groups, even if the nested groups are
defined before the examples.

Running tests in a defined order can hide order-dependent tests or other
erratic behavior. Setting the order to `random` randomizes the order of
groups and examples within the groups. It will also display the random
seed which can be used to duplicate a particular execution order in the
event of a failure.