(vim) Automatically show list of search results
When `:Grep` and `:LGrep` have completed, their respective window must
be opened manually in order to view the search results. We can instead
use two `autocmd`s to open the window automatically -- one to trigger
`:cwindow` when `:cgetexpr` is executed and one to trigger `:lwindow`
when `:lgetexpr` is executed. `:cwindow` and `:lwindow` open their
respective window only if there are valid entries.