(psql) Set ON_ERROR_ROLLBACK interactive
The `ON_ERROR_ROLLBACK` setting affects how `psql` handles errors. The
default is `off`, which causes an error in a transaction block to abort
the entire transaction. When set to `on`, an error in a transaction
block is ignored and the transaction continues. When set to
`interactive`, an error in a transaction is ignored only in interactive
sessions, and not when reading script files.