(psql) Auto-format based on output
SQL output can have a huge variance in the number of columns for
display. The format of the output can be toggled to be expanded
or compressed by setting `\x on` or `\x off`, respectively.

Additionally, setting `\x auto` tells Postgres to automatically detect
which format is most ideal for the output based on the width of the
result set and the width of the terminal.