(repo) Exclude repository files in ripgrep
Ripgrep automatically ignores files and directories that match patterns
in `.gitignore`. It does this, however, only when it exists inside of
the repository.

Since this repository is set up as a bare repository, where the working
tree differs from the git directory, ripgrep does not exclude patterns
in `.gitignore`. Symlinking `.gitignore` to `.ignore` gets around this.