(fish) Wrap git to account for dotfiles bare repo
The dotfiles repository is a bare repository with the working tree
directory separate from the git directory. This requires that the
`--git-dir` and `--work-tree` options be provided to `git`.

Remembering to specify these options each time is verbose and prone to
error. This adds a fish function that wraps `git` to supply the needed
options when called from with the dotfiles repository (i.e., $HOME).