(weechat) Update GnuTLS CA file location
Weechat is configured to connect to IRC networks over SSL and identify
me using an SSL certificate. It checks that the SSL connection and SSL
certificate are fully trusted and refuses to connect when it's not.

```
=!= gnutls: peer's certificate is NOT trusted
=!= gnutls: peer's certificate issuer is unknown
=!= irc: TLS handshake failed
=!= irc: error: Error in the certificate.
```

Weechat uses the certificate authorities (CA) file at `gnutls_ca_file`
for this check. `openssl` used to bootstrap this file and places it at
`/usr/local/etc/openssl/cert.pem`. Now, however, `openssl` is an alias
for `openssl@1.1`, which bootstrap this file and places it at
`/usr/local/etc/openssl@1.1/cert.pem`.