(weechat) Auto refresh all options on fset buffer
When an option is modified, the fset buffer is refreshed. This can have
undesirable side-effects (e.g., CPU spike) when a trigger exists that
modifies options at a frequent rate (e.g., every second).

An option was introduced in Weechat 3.0 to set a command separated list
of options to automatically refresh on the fset buffer. "*", the default
and the recommended value, refreshes all options. "!" also exists as a
negative value to prevent an option from being refreshed.

https://github.com/weechat/weechat/issues/1553
https://weechat.org/files/changelog/ChangeLog-3.0.html#v3.0