(weechat) Enable required opt-in for downloads
Once `/script` has been run and plugins.xml.gz exists, an automatic
refresh would occur that caused Weechat to phone-home a request for
updated information.

This reveals information about the user (e.g., ISP, IP) and may not
always be desirable. Furthermore, it was done without consent and,
often times, awareness.

An option was introduced in Weechat 3.0 to disable this feature by
default and require opt-in of any download. Downloads can be enabled as
needed with `/set script.scripts.download_enabled on`.

https://github.com/weechat/weechat/issues/1548
https://weechat.org/files/changelog/ChangeLog-3.0.html#v3.0