(vim) Switch to vim-bettergrep
The `vim-bettergrep` plugin provides lightweight and enhanced versions
of the `:grep` family of commands. Its aim is to closely imitate the
original commands, and to only provide a minimal set of enhancements
based on RomainL's instant grep + quickfix.

Initial testing shows that this plugin handles multiple word searches
and command line options better than the previous implementation, and
was the main motivation behind switching.

https://gist.github.com/romainl/56f0c28ef953ffc157f36cc495947ab3