(vim) Reduce height of vim-test terminal window
By default, the `vimterminal` strategy will open the new window with
`botright`, which positions the window at the bottom using the full
width of the Vim window and half of the Vim window height.

As that's more screen real estate than is needed for test output, create
instead a new window with the same positioning as the default, but with
fewer (20) rows.