(vim) Use indent style to split method arguments
When splitting method arguments, the hanging style for method argument
alignment is used by default. This keeps the first method argument on
the same line as the method name, and aligns subsequent arguments
directly beneath.

The hanging style is an accepted and commonly used style but isn't my
preference. Instead, when splitting method arguments, use the normal
indent style which starts the arguments list on the next line and
follows standard indentation rules.