(vim) Split Ruby method arguments and options
By default, splitting method arguments will always split the options, if
there are any, and only split the positional arguments, if there are no
options. This behavior often times leads to inconsistent formatting.

To maintain consistent formatting across various argument types, split
across all argument types, if the cursor is on the arguments, but only
split the options if the cursor is on the options.