(kitty) Use left option key as alt key
When set to `no`, kitty will use the macOS native `option+key` behavior
and insert a unicode character. This breaks any keyboard shortcuts that
depending on an `alt` or `meta` key combination (e.g., readline).

Setting this to `left` treats the left option key as if it were an alt
key, and keeps the right option key as is. This allows using keyboard
shortcuts that depend on having the `alt` key available, also without
sacrificing the ability to write unicode characters.