(vim) Remove vim-projectionist
There is an issue with recent versions of vim (>8.2.413, 5e94a29) that
causes the cursor to jump to the first line of the current buffer when a
hidden terminal buffer is created and a `User` event is executed after
`BufFilePost`.

`vim-projectionist` triggers said `User` event (`ProjectionistDetect`)
after `BufFilePost`. The cursor jumping to the first line of the buffer
has made keep track of location when navigating multiple buffers
difficult and confusing. As such, this removes `vim-projectionist`
until a fix is available.