(vim) Rebalance windows when Vim window is resized
Opening a tmux pane shifts other panes to share the terminal window
space. When one of those panes contains Vim, the main Vim window is
resized. Any open windows are not automatically resized, however,
causing the windows to sometimes not to be visible. This addresses the
issue by rebalancing all open windows whenever the main Vim window is
resized.