(vim) Install vim-zoom plugin
vim-zoom is a plugin to toggle the zoom of the current window. Usually
having multiple windows open at a given time, it's handy to have the
ability to temporarily focus in on a particular window.

Having fewer windows open at any given time, and using `:Buffer` to
navigate efficiently between windows may be a better option to work
towards and that would make this plugin unnecessary.