(fish) Use fd as default fzf command
While using `rg` as the default `fzf` command works just fine, `fd` is
purpose built for listing files, which is what we want for `fzf`, and
has similar performance characteristics.

This configures `fzf`'s in the shell so that the configuration applies
to anywhere that `fd` is run, and not only from within Vim.