(vim) Configure standard as code formatter
This adds `standard` as a formatter for `diagnostic-languageserver`. It
runs `standard` with the `--fix` flag to fix any violations found in the
file automatically.

The `--force-exclusion` option excludes files passed to `standard` as an
argument when they have been `Exclude`d in configuration. This prevents
linting of files that have been explicitly excluded.

The formatter only runs when one of the `requiredFiles` is present. The
markers defined in `rootPatterns` are used to traverse upwards from the
current directory to the root directory when searching for the required
files.

The formatter can be triggered with `:call CocAction('format')`.