(vim) Configure standard as code linter
This adds `standard` as a linter for `diagnostic-languageserver`. It
runs `standard` to gather information on any violations which are then
parsed and displayed in the editor.

The `--force-exclusion` option excludes files passed to `standard` as an
argument when they have been `Exclude`d in configuration. This prevents
linting of files that have been explicitly excluded.

The linter runs automatically, but only when one of the `requiredFiles`
is present. The markers defined in `rootPatterns` are used to traverse
upwards from the current directory to the root directory when searching
for the required files.