(vim) Enable format on type in CoC
`coc-tsserver` runs formatting on type by default but it does not take
effect unless also enabled in CoC. [`formatOnTypeFiles`] doesn't appear
to be necessary for `javascript` and `typescript` file types as the
formatting for these file types is enabled by `coc-tsserver`.

The [`formatOnTypeFiles`] option could be used to enable formatting for
these file types explicitly, or to list additional file types for on
type formatting. Leaving the option as its default for now, until doing
so becomes necessary.

[`formatOnTypeFiles`]: https://github.com/neoclide/coc.nvim/blob/master/data/schema.json#L1089-L1096