(nvim) Rebalance windows when window is resized
Opening a tmux pane shifts other panes to share the terminal window
space. When one of those panes contains neovim, the main neovim window
is resized. Any open windows are not automatically resized, however,
causing the windows to sometimes not to be visible. This addresses the
issue by rebalancing all open windows whenever the main neovim window is
resized.