(nvim) Configure Treesitter
nvim-treesitter wraps the Neovim treesitter API to provide functionality
for highlighting, incremental selection, and more. All available modules
are disabled by default, and only highlighting is enabled.

The `additional_vim_regex_highlighting` option is enabled to run syntax
highlighting and treesitter at the same time. This is necessary for the
`vim-endwise` plugin to function, but may slow down the editor or cause
duplicate highlights.

Additionally, all maintained language parsers are installed, though many
are for languages that aren't used. Other than the time needed to first
install the parsers, this shouldn't be an issue and keeps us up-to-date
as new parsers are made available.