(nvim) Configure complete LSP setup
Neovim supports the Language Server Protocol (LSP) and provides a common
interface for common features, such as go-to-definition, completion,
etc. Neovim provides an interface, but is less out-of-the-box ready and
requires additional configuration.

* Configure nvim-lspconfig

Provides a collection of common configurations for Neovim's built-in
language server client, and allows for configuring, launching, and
initializing language servers installed on the system.

* Configure nvim-lsp-installer

Provides easy installation and management of languages servers. It is
optimized for fast startup times, provides a simple GUI, and supports
installing custom versions (e.g., `rust_analyzer@nightly`).

* Configure lspkind

Provides vscode-like pictograms for Neovim's built-in LSP. It requires
that the vscode-codicons font is installed to correctly display the
icons.

* Configure lsp_signature

Provides function signature details (e.g., argument list, return value)
in a pop up window for the function under the cursor.

* Configure nvim-cmp and sources

Provides a completion engine for Neovim. Completion sources are sourced
from external repositories and configured.