(nvim) Add Zen Mode for a distraction-free editor
For the times where focus on a single thing is needed, disable all of
the extras in Neovim and Tmux, and bump the font size. The ability to
increase the font size requires that Kitty listen for commands on the
specified socket, and that Tmux also has the correct socket set for
communication with Kitty.