(fish) Wrap nvim to handle dotfiles bare repo
The dotfiles repository is a bare repository with the working tree
directory separate from the git directory. This requires that the
`--git-dir` and `--work-tree` options be provided to `git`.

Remembering to specify these options each time is verbose and prone to
error. This adds a fish function that wraps `nvim` to supply the needed
options when called from

NOTE: This is to make vim-fugitive work with the dotfiles setup, and
requires that `git config core.worktree` be set to `$HOME`.