(nvim) Use prettierd instead of prettier
Similar to `eslint_d` providing a faster alternative to `eslint`, using
`prettierd` provides a faster alternative to `prettier`, buy running a
server in the background.

The formatter is configured to only run when a `prettier` configuration
file exists, and will use a local version of `prettierd`, if installed,
before falling back to the globally installed option.

`npm install -g @fsouza/prettierd`