(nvim) Set the text width for the Ruby filetype
Setting `textwidth` determines the maximum width of text, and indirectly
determines the value of `colorcolumn` which is configured to highlight
the column after the column set for `textwidth`.

Limiting code to 80 characters is easier to read and my target when
coding. Configuring Vim to be aware of it, particularly having
`colorcolumn` enabled, helps stick to it.