(nvim) Correctly append to `iskeyword` option
The previous syntax used string concatenation (+) to add the - character
to the `iskeyword` list, but didn't account for the comma-separated list
expected by `iskeyword`. The `+` operator can be used, but requires that
a table argument is provided to correctly append to the option list
(e.g., `vim.bo.iskeyword + {"-"}`).

Alternatively, this uses the `:append` method to accomplish the same end
result, and uses `vim.opt_local` to replicate the behavior of `setlocal`
which sets the option only local to the buffer.