(nvim) Move plugin files out of `runtimepath`
Neovim includes the plugin directory in the `runtimepath`, which means
that any files included in this directory are sourced at runtime. This
has the disadvantage of loading more at start up than is necessary.

For future changes to lazy load plugins for faster startup times, this
instead moves the files to a directory outside of the `runtimepath` to
allow them to then be required manually.