(nvim) Use explicit list of language parsers
With the increasing number of parsers added to `nvim-treesitter`, having
a maintainer listed no longer entails any stability guarantee implied by
this label. The plugin authors deprecated this option in favor of adding
an explicit list of parsers to install according to need. Installing all
parsers is an option, but is not recommended.