(nvim) Switch to main branch of lspactions
lspactions [removed] the [deprecated] astronauta.nvim as a dependency in
favor of the built-in `vim.keymap` added to Neovim. This required Neovim
nightly for the new functionality or the branch of lspactions compatible
with Neovim 0.6 for previous functionality. With 0.7 released upgrade to
the latest Neovim and switch to lspactions' main branch.

[removed]: https://github.com/RishabhRD/lspactions/pull/8
[deprecated]: https://github.com/tjdevries/astronauta.nvim/commit/c8e33813c92ffcc9ba2edaf9124db1f3f17e3500