(nvim) Add which-key plugin for key bindings
Having minimal key bindings in general and even fewer that are difficult
to remember, this plugin hasn't felt necessary. Adding more key bindings
that are infrequently used and greater in number (e.g., for note taking)
the reminder now sometimes comes in handy.

Setting `ignore_missing = true` ignores key bindings that were not added
explicitly to which-key, which allows opting-in to the reminder behavior
and avoids the reminder pop up unnecessarily.