(nvim) Fix running visually selected SQL
vim-dadbod has preliminary map support and allows defining key mappings
to execute SQL using normal mode and visual mode. This is handy for SQL
under the cursor or a SQL file. To be able to run the visually selected
SQL, this fixes one of the key mappings to be an `xmap` key mapping.