(nvim) Use eslint LSP server instead of eslint_d
Using `eslint_d` provided a faster alternative to `eslint` by spawning a
server in the background. It was the preferred option, but it was known
to be memory intensive and to cause machines to come to a crawl.

Microsoft's eslint language server now provides a solid alternative that
is has the speed expected of an LSP, without the memory issues caused by
`eslint_d`