(nvim) Use internal formatter for gitcommit files
[Changes in Neovim 0.8] default `formatexpr` to `vim.lsp.formatexpr()`
when `formatexpr` and `formatprg` are not set, and the language server
attached to the buffer supports the `documentRangeFormattingProvider`
specification.

`formatexpr` and `formatprg` are unset, and the null-ls set up attaches
diagnostic and code action sources that may be why `gitcommit` buffers
show support for `documentRangeFormattingProvider`.

Using `vim.lsp.formatexpr()` with the attached language servers breaks
`gq`, which is useful for reformatting commit messages. Until resolved
another way, set `formatexpr` explicitly to use the internal formatter.

[Changes in Neovim 0.8]: https://github.com/neovim/neovim/blob/9b4cab012662514af6fda3648d544633e1d73d4b/runtime/lua/vim/lsp.lua#L1105-L1112