(nvim) Refactor LSP + null-ls setup
This takes a pass at general clean up of duplication around language
server setup, as well as uses `client.supports_method` to set up the
commands and keymaps supported by the language server, and not other
commands and keymaps, as done before. The exception being diagnostic
commands and keymaps, which are always set up because of an issue w/
null-ls not showing support for `textDocument/publishDiagnostics`.