(nvim) Trigger `WhichKey` on `<leader>`
`vim-rails` extends `gf` with go to file support for Rails codebases but
does so if a `gf` map doesn't already exist. It checks for this with the
`mapcheck('gf', ...)` command, which matches mappings that match the
start of `gf` (.i.e., `gf` matches on `g`).

`WhichKey` defines mappings for presets including mappings for commands
prefixed with "g". This causes an issue because the presence of the `g`
map causes `vim-rails` not to extend `gf` with Rails support.

Options exist to disable mappings for `g` and other presets, but doing
this doesn't remove the `g` map. Since custom mappings all start with
`<leader>` and extending `gf`'s more important than preset mappings,
trigger `WhichKey` specifically for `<leader>` mappings.