(asdf, brew, nvim) Write better with Vale
As mentioned [before] writing's hard. This moves from separate tools to
Vale as a single source, with packages for the prior tools and more. It
makes the following changes:

Use Homebrew to install Vale: This journey starts with installing Vale.
To keep it up-to-date and install it on new machines, manage the tool
with Homebrew.

Use Vale packages for prior tools: Remove `proselint` and `writegood` as
default packages. Let Vale manage these instead. Install and update them
with`vale sync`.

Swap null-ls from `proselint` and `write_good` to `vale`. Integrate Vale
with Neovim instead using `null-ls`' built-in Vale configuration to show
diagnostics.

Control what Vale lints, how it lints, and where it lints. Use Microsoft
style guide as a base, and the `proselint`, `write-good`, `Readability`,
`alex`, and `JobLint` packages.

Ignore vale's synced styles and packages. Vale installs its packages to
~/.vale/styles. Exclude these files from this repository.

[before]: https://github.com/emilford/dotfiles/tree/e2bc108552af40a6e7d6bda698705721ba434f64