(nvim) Keep CTRL-T behavior in incremental search
When incremental searching CTRL-G and CTRL-T jump to the next and prior
match respectively before completing the search. This is useful to look
ahead or behind to inform the search.

readline.vim overrides CTRL-T to transpose the characters before the
cursor. This isn't useful for my needs. To preserve the default
behavior, this disables the readline.vim mapping.