(nvim) Disable 3rd-party Lua LSP checks
Upgrading lua-language-server resulted in a prompt to configure the lua
work environment. No idea what this feature does, which for now means
disable it until understood and needed.

```
Do you need to configure your work environment as `luassert`?
Request Actions:
1. Apply and modify settings
2. Apply but do not modify settings
3. Don't show again
Type number and <Enter> or click with the mouse (q or empty cancels):
```