(nvim) Add keymap for Telescope `find_files`
Telescope's `git_files` fuzzy searches the output of `git ls-files` and
is relative to the repository root directory. This means when viewing a
gem with `bundle open`, that fuzzy finding files with `git_files` shows
the root files instead of files relative to the `vendor/bundle` path.

For this and similar cases, this adds a fallback keymap to `find_files`
which lists files relative to the current directory. There may be the
opportunity to add smarts around a single keymap. This works for now.