(nvim) Don't select completion option on Enter
With the prior setting (i.e., `select = true`), hitting enter with the
completion window open would select the first completion option, which
would expand it. This would often lead to undesired behavior, when not
selecting a completion option and instead finishing typing and hitting
enter. This change resolves the issue and leaves `c-y` for to opt-into
the prior behavior.