(fish, ngrok) Add ngrok configuration from XDG
This started adding the `custom_ui_background` configuration to ngrok to
change the default background that's inconsistent from other window. The
default configuration file on MacOS is outside of the home directory and
also includes sensitive data. This allows for separate configuration and
can exist securely in the dotfiles repository.