(nvim) Use minimal drop down Telescope finders
This switches to the built-in dropdown theme with a few customizations
to have a minimal finder window for fuzzy finding files. Fuzzy finding
help documentation still uses the horizontal layout and preview.

This also removes the mapping to toggle the preview window. Seemed like
it could be handy, but it's not useable with the center layout—which is
what the dropdown theme is built on—and wasn't used much anyway.