(tmux) Use fish shell when starting tmux
This change resolves a tricky issue where the asdf directories were no
ahead of other directories (e.g., `/usr/bin`) in `$PATH`, breaking the
use of asdf to manage and select package versions.

The issue stems from [this PR] that improved how asdf adds directories
to `$PATH`, but that surfaced [an issue] found in shells other than
fish shell.

Noticing that this issue happened when in a tmux session and otherwise
worked fine led to changing the default command to fish which resolved
the issue. Though unclear why `$SHELL` doesn't work, per the docs.

[this PR]: https://www.github.com/asdf-vm/asdf/pull/1709
[an issue]: https://www.github.com/asdf-vm/asdf/issues/1550