(nvim) Add initial DAP debug setup + Ruby adapter
Sets up Neovim as a DAP (Debug Adapter Protocol) client. Configures DAP
UI with watches, breakpoints, and the stackframe, and adds a key map to
launch the REPL in a floating window as needed.

The Ruby setup uses a plugin to configure the adapter but overrides its
configurations to make them filetype-specific and to limit them to ones
most useful: test current file, debug current file, debug w/ server.