(fish, kitty, nvim) Explore dropping tmux
Switching to Neotest meant having one less need for tmux. In the past
tmux has caused problems and still has its downsides, but features it
offers made the trade-offs worth it.

This experiments with Kitty as a replacement. The remaining pieces of
the switch is effective switching between open sessions and re-naming
sessions. The current solution isn't ready for prime time.

Kitty:

- Configures the tab bar to match prior look and feel
- Configures the light and dark themes to match prior look
- Configures layouts to match prior window behavior
- Configures key maps to match prior behaviors

Neovim:

- Adds vim-kitty-navigator for navigation between nvim and Kitty

Fish:

- Removes setting tmux's theme from theme switcher