(nvim) Set `textwidth` for `eruby` files
`eruby` templates tend to have long lines because views tend not to get
the same care on line length in Rails applications as other types files
for whatever reason. Setting a `textwidth` enables `cursorcolumn` which
shows a marker for a reasonable line length for views, which is helpful
for refactoring the long lines.