(nvim) Set up vim-test as a backup test runner
Still trying out neotest but sometimes running tests in a shell using
the CLI test runner makes seeing the test run easier. To support that
this enables the plugin and reconfigures the key maps to use "T"

Not using Tmux, this also updates the test strategy to use Kitty, and
adds a keyboard shortcut to close the vim-test output window.