(nvim) Run tests using a modified WezTerm strategy
Toggleterm had a couple of weird behaviors. Navigating in and out of the
terminal felt clunky, and zooming the terminal buffer didn't unwrap text
which defeated the purpose of zooming in.

This wraps vim-test's original WezTerm strategy to experiment to resolve
the current issue where after closing the WezTerm pane, it won't re-open
on following test runs. Clearing the `test_id` seems to fix the issue.